// src/frontend/src/app/enums/enums.ts
export enum Type {
  Lost = 'LOST',
  Seen = 'SEEN',
  Found = 'FOUND',
  Unknown = 'UNKNOWN',
}

export enum Animal {
  Dog = 'DOG',
  Cat = 'CAT',
  Other = 'OTHER',
  Unknown = 'UNKNOWN',
}

export enum Size {
  Small = 'SMALL',
  Medium = 'MEDIUM',
  Large = 'LARGE',
  Unknown = 'UNKNOWN',
}

export enum Gender {
  Male = 'MALE',
  Female = 'FEMALE',
  Unknown = 'UNKNOWN',
}

export enum Age {
  Young = 'YOUNG',
  Adult = 'ADULT',
  Senior = 'SENIOR',
  Unknown = 'UNKNOWN',
}

export enum Status {
  Active = 'ACTIVE',
  Updated = 'UPDATED',
  Inactive = 'DELETED',
  Unknown = 'FOUND',
  Blocked = 'BLOCKED'
}